backend/requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
openai
google-genai
python-dotenv
pytest
